Implement automatic deployment to sysc4907.macdonell.net
This changeset adds the ability to deployment to macdonell.net via
capistrano. Additionally, it also handles running multiple processes
for the eventual addition of delayed job workers and periodic tasks.

To run the application locally, run `foreman start`.

Remotely, the application is run by supervisord. This means that we
can add more services just by editing the Profile and redeploying.